Used npm instead of pnpm
Here is why:
I added the _config.yml file to include node_modules when building the website. (Why?) because node_modules is excluded by default when building the website with Jekyllrb. It turns out that Jekyllrb also excludes files that start with a dot (.), and one of the needed files is in a `.pnpm` folder. I tried to include it too in the _config.yml, but it didn't work, so I reverted to npm which does not put the packages in a dot file.